add new package.json and rename files
