--- src/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "Class-stack",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./moduls/Class-Stack.js",
-  "scripts": {
-    "test": "mocha ./moduls/Class-Stack.test.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "chai": "^4.3.4"
-  },
-  "devDependencies": {
-    "mocha": "^9.1.3"
-  }
-}
